ml-engine docker container & structure figured out
