+++ .gitignore
@@ -4,3 +4,4 @@ mosquitto/data/
 influxdb-data/
 mongo-data/
 .sock
+__pycache__
