+++ ml-engine/handler_imports.py
@@ -1,0 +1,1 @@
+import handlers.test_handlers
