+++ ml-engine/requirements.txt
@@ -1,0 +1,1 @@
+json-rpc==1.15.0
